+++ src/main.cpp
@@ -246,6 +246,9 @@ int main(int argc, char** argv) {
     } else if (std::strcmp(arg, "--freeze") == 0) {
       options.freezePlayer = true;
       wantsWorld = true;
+    } else if (std::strcmp(arg, "--creative") == 0) {
+      options.creativeWorld = true;
+      wantsWorld = true;
     } else if (std::strcmp(arg, "--debug-lines") == 0) {
       options.debugLines = true;
     } else if (std::strcmp(arg, "--debug-view") == 0) {